fix: Implement explicit waits in loginWithCredentials
Refactored the loginWithCredentials method in the EchoDownloader class
to use Selenium's WebDriverWait for explicit waits.

This change the script reliably waits for the presence and interactability
of the username and passowrd input fields, as well as the login button.
This imporovement addresses issues related to elements not being loaded in time
during the login process.